requirements.txt
pillow==10.3.0
transformers==4.41.2
sentence_transformers==2.7.0
pandas==2.2.2
matplotlib==3.9.0
streamlit==1.36.0
streamlit_feedback==0.1.3
streamlit_option_menu==0.3.13
google-generativeai>=0.5.2
python-dotenv==1.0.1
langchain==0.2.7
pypdf==4.2.0
chromadb==0.5.3
faiss-cpu==1.8.0
langchain_google_genai
langchain_openai==0.1.14
langchain-community==0.2.7
langchain_huggingface==0.0.3
langchain_groq==0.1.6
jq==1.7.0
numpy==1.26.4
matplotlib==3.9.0
s3fs==2024.6.1
sagemaker==2.224.4
llama_index_finetuning==0.1.10
llama-index==0.10.53
llama-index-embeddings-huggingface==0.2.2
huggingface_hub==0.23.4
scikit-learn==1.5.1
boto3==1.34.131
-e .